Line done in OOP and TDD
This is not the best commit message one could ever think. However, the purpose of what I am doing now is to re-write the introductory programming exercises in an OOP and TDD way. So, in the circle (Red-Green-Blue) and then commit, I skip the commit part a lot of times, because these are very simple exercises.